+++ Cargo.toml
@@ -1,6 +1,8 @@
 [workspace]
 members = [
-  "node",
+  "node/main",
+  "node/staking",
+  "node/parachain",
   "pallets/*",
   "primitives",
   "runtimes/*",
