+++ node/parachain/src/chain_spec.rs
@@ -1,0 +1,290 @@
+/*
+ * This file is part of the Nodle Chain distributed at https://github.com/NodleCode/chain
+ * Copyright (C) 2022  Nodle International
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+use cumulus_primitives_core::ParaId;
+use primitives::{AccountId, Balance, Signature};
+use runtime_eden::{
+    constants::NODL, AuraId, BalancesConfig, GenesisConfig, ParachainInfoConfig, SessionConfig,
+    SessionKeys, SudoConfig, SystemConfig, ValidatorsSetConfig, WASM_BINARY,
+};
+use sc_chain_spec::{ChainSpecExtension, ChainSpecGroup};
+use sc_service::ChainType;
+use serde::{Deserialize, Serialize};
+use sp_core::{sr25519, Pair, Public};
+use sp_runtime::traits::{IdentifyAccount, Verify};
+
+/// Specialized `ChainSpec` for the normal parachain runtime.
+pub type ChainSpec = sc_service::GenericChainSpec<GenesisConfig, Extensions>;
+
+/// Helper function to generate a crypto pair from seed
+pub fn get_public_from_seed<TPublic: Public>(seed: &str) -> <TPublic::Pair as Pair>::Public {
+    TPublic::Pair::from_string(&format!("//{}", seed), None)
+        .expect("static values are valid; qed")
+        .public()
+}
+
+/// The extensions for the [`ChainSpec`].
+#[derive(Debug, Clone, PartialEq, Serialize, Deserialize, ChainSpecGroup, ChainSpecExtension)]
+#[serde(rename_all = "camelCase")]
+pub struct Extensions {
+    /// The relay chain of the Parachain.
+    pub relay_chain: String,
+    /// The id of the Parachain.
+    pub para_id: u32,
+}
+
+impl Extensions {
+    /// Try to get the extension from the given `ChainSpec`.
+    pub fn try_get(chain_spec: &dyn sc_service::ChainSpec) -> Option<&Self> {
+        sc_chain_spec::get_extension(chain_spec.extensions())
+    }
+}
+
+type AccountPublic = <Signature as Verify>::Signer;
+
+/// Generate collator keys from seed.
+///
+/// This function's return type must always match the session keys of the chain in tuple format.
+pub fn get_collator_keys_from_seed(seed: &str) -> AuraId {
+    get_public_from_seed::<AuraId>(seed)
+}
+
+/// Helper function to generate an account ID from seed
+pub fn get_account_id_from_seed<TPublic: Public>(seed: &str) -> AccountId
+where
+    AccountPublic: From<<TPublic::Pair as Pair>::Public>,
+{
+    AccountPublic::from(get_public_from_seed::<TPublic>(seed)).into_account()
+}
+
+/// Generate the session keys from individual elements.
+///
+/// The input must be a tuple of individual keys (a single arg for now since we have just one key).
+pub fn eden_session_keys(keys: AuraId) -> SessionKeys {
+    SessionKeys { aura: keys }
+}
+
+/// Helper function to create GenesisConfig for testing
+fn eden_testnet_genesis(
+    root_key: AccountId,
+    collators: Vec<(AccountId, AuraId)>,
+    endowed_accounts: Option<Vec<AccountId>>,
+    id: ParaId,
+) -> GenesisConfig {
+    let endowed_accounts: Vec<AccountId> = endowed_accounts.unwrap_or_else(|| {
+        vec![
+            get_account_id_from_seed::<sr25519::Public>("Alice"),
+            get_account_id_from_seed::<sr25519::Public>("Bob"),
+            get_account_id_from_seed::<sr25519::Public>("Charlie"),
+            get_account_id_from_seed::<sr25519::Public>("Dave"),
+            get_account_id_from_seed::<sr25519::Public>("Eve"),
+            get_account_id_from_seed::<sr25519::Public>("Alice//stash"),
+            get_account_id_from_seed::<sr25519::Public>("Bob//stash"),
+            get_account_id_from_seed::<sr25519::Public>("Charlie//stash"),
+            get_account_id_from_seed::<sr25519::Public>("Dave//stash"),
+            get_account_id_from_seed::<sr25519::Public>("Eve//stash"),
+        ]
+    });
+
+    const ENDOWMENT: Balance = 10_000 * NODL;
+
+    GenesisConfig {
+        // Core
+        system: SystemConfig {
+            code: WASM_BINARY
+                .expect("WASM binary was not build, please build it!")
+                .to_vec(),
+        },
+        balances: BalancesConfig {
+            balances: endowed_accounts
+                .iter()
+                .cloned()
+                .map(|k| (k, ENDOWMENT))
+                .collect(),
+        },
+        sudo: SudoConfig {
+            key: Some(root_key),
+        },
+        vesting: Default::default(),
+
+        // Consensus
+        validators_set: ValidatorsSetConfig {
+            members: collators.iter().map(|x| x.0.clone()).collect::<Vec<_>>(),
+            phantom: Default::default(),
+        },
+        session: SessionConfig {
+            keys: collators
+                .into_iter()
+                .map(|(acc, aura)| {
+                    (
+                        acc.clone(),             // account id
+                        acc,                     // validator id
+                        eden_session_keys(aura), // session keys
+                    )
+                })
+                .collect(),
+        },
+        aura: Default::default(),
+        aura_ext: Default::default(),
+        parachain_system: Default::default(),
+        parachain_info: ParachainInfoConfig { parachain_id: id },
+
+        // Governance
+        company_reserve: Default::default(),
+        international_reserve: Default::default(),
+        usa_reserve: Default::default(),
+
+        // Allocations
+        allocations_oracles: Default::default(),
+    }
+}
+
+fn development_config_genesis(id: ParaId) -> GenesisConfig {
+    eden_testnet_genesis(
+        get_account_id_from_seed::<sr25519::Public>("Alice"),
+        vec![
+            (
+                get_account_id_from_seed::<sr25519::Public>("Alice"),
+                get_collator_keys_from_seed("Alice"),
+            ),
+            (
+                get_account_id_from_seed::<sr25519::Public>("Bob"),
+                get_collator_keys_from_seed("Bob"),
+            ),
+        ],
+        None,
+        id.into(),
+    )
+}
+
+pub fn development_config(id: ParaId) -> ChainSpec {
+    // Give your base currency a unit name and decimal places
+    let mut properties = sc_chain_spec::Properties::new();
+    properties.insert("tokenSymbol".into(), "NODL".into());
+    properties.insert("tokenDecimals".into(), 11.into());
+    properties.insert("ss58Format".into(), 42.into());
+
+    ChainSpec::from_genesis(
+        // Name
+        "Parachain Eden Development",
+        // ID
+        "para_eden_dev",
+        ChainType::Development,
+        move || development_config_genesis(id),
+        Vec::new(),
+        None,
+        None,
+        None,
+        None,
+        Extensions {
+            relay_chain: "rococo-local".into(), // You MUST set this to the correct network!
+            para_id: id.into(),
+        },
+    )
+}
+
+fn local_config_genesis(id: ParaId) -> GenesisConfig {
+    eden_testnet_genesis(
+        get_account_id_from_seed::<sr25519::Public>("Alice"),
+        vec![
+            (
+                get_account_id_from_seed::<sr25519::Public>("Alice"),
+                get_collator_keys_from_seed("Alice"),
+            ),
+            (
+                get_account_id_from_seed::<sr25519::Public>("Bob"),
+                get_collator_keys_from_seed("Bob"),
+            ),
+            (
+                get_account_id_from_seed::<sr25519::Public>("Charlie"),
+                get_collator_keys_from_seed("Charlie"),
+            ),
+            (
+                get_account_id_from_seed::<sr25519::Public>("Dave"),
+                get_collator_keys_from_seed("Dave"),
+            ),
+            (
+                get_account_id_from_seed::<sr25519::Public>("Eve"),
+                get_collator_keys_from_seed("Eve"),
+            ),
+        ],
+        None,
+        id.into(),
+    )
+}
+
+pub fn local_testnet_config(id: ParaId) -> ChainSpec {
+    // Give your base currency a unit name and decimal places
+    let mut properties = sc_chain_spec::Properties::new();
+    properties.insert("tokenSymbol".into(), "NODL".into());
+    properties.insert("tokenDecimals".into(), 11.into());
+    properties.insert("ss58Format".into(), 42.into());
+
+    ChainSpec::from_genesis(
+        // Name
+        "Eden Local Testnet",
+        // ID
+        "para_eden_local",
+        ChainType::Local,
+        move || local_config_genesis(id),
+        // Bootnodes
+        Vec::new(),
+        // Telemetry
+        None,
+        // Protocol ID
+        Some("eden-local"),
+        // Fork ID
+        None,
+        // Properties
+        Some(properties),
+        // Extensions
+        Extensions {
+            relay_chain: "westend".into(), // You MUST set this to the correct network!
+            para_id: id.into(),
+        },
+    )
+}
+
+pub fn main_config() -> ChainSpec {
+    ChainSpec::from_json_bytes(&include_bytes!("../../res/eden.json")[..]).unwrap()
+}
+
+#[cfg(test)]
+pub(crate) mod tests {
+    use super::*;
+    use sp_runtime::BuildStorage;
+
+    #[test]
+    fn test_create_development_chain_spec() {
+        development_config(ParaId::from(1000u32))
+            .build_storage()
+            .unwrap();
+    }
+
+    #[test]
+    fn test_create_local_chain_spec() {
+        local_testnet_config(ParaId::from(1000u32))
+            .build_storage()
+            .unwrap();
+    }
+
+    #[test]
+    fn test_create_main_spec() {
+        main_config().build_storage().unwrap();
+    }
+}
