+++ node/res/arcadia.json
@@ -25,8 +25,6 @@
     "tokenDecimals": 11,
     "tokenSymbol": "aNODL"
   },
-  "relayChain": "",
-  "paraId": 0,
   "consensusEngine": null,
   "genesis": {
     "raw": {
