+++ node/res/staking.full.json
@@ -19,8 +19,6 @@
     "tokenDecimals": 11,
     "tokenSymbol": "sNODL"
   },
-  "relayChain": "",
-  "paraId": 0,
   "consensusEngine": null,
   "codeSubstitutes": {},
   "genesis": {
