--- node/src/cli.rs
@@ -1,130 +1,0 @@
-/*
- * This file is part of the Nodle Chain distributed at https://github.com/NodleCode/chain
- * Copyright (C) 2022  Nodle International
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
- */
-use clap::Parser;
-use sc_cli::{KeySubcommand, SignCmd, VanityCmd, VerifyCmd};
-use std::path::PathBuf;
-
-/// An overarching CLI command definition.
-#[derive(Debug, Parser)]
-#[clap(
-    propagate_version = true,
-    args_conflicts_with_subcommands = true,
-    subcommand_negates_reqs = true
-)]
-pub struct Cli {
-    /// Possible subcommand with parameters.
-    #[clap(subcommand)]
-    pub subcommand: Option<Subcommand>,
-
-    #[clap(flatten)]
-    pub run: cumulus_client_cli::RunCmd,
-
-    /// Relay chain arguments
-    #[clap(raw = true)]
-    pub relaychain_args: Vec<String>,
-}
-
-/// Command for exporting the genesis state of the parachain
-#[derive(Debug, Parser)]
-pub struct ExportGenesisStateCommand {
-    /// Output file name or stdout if unspecified.
-    #[clap(parse(from_os_str))]
-    pub output: Option<PathBuf>,
-
-    /// Write output in binary. Default is to write in hex.
-    #[clap(short, long)]
-    pub raw: bool,
-
-    /// The name of the chain for that the genesis state should be exported.
-    #[clap(long)]
-    pub chain: Option<String>,
-}
-
-/// Command for exporting the genesis wasm file.
-#[derive(Debug, Parser)]
-pub struct ExportGenesisWasmCommand {
-    /// Output file name or stdout if unspecified.
-    #[clap(parse(from_os_str))]
-    pub output: Option<PathBuf>,
-
-    /// Write output in binary. Default is to write in hex.
-    #[clap(short, long)]
-    pub raw: bool,
-
-    /// The name of the chain for that the genesis wasm file should be exported.
-    #[clap(long)]
-    pub chain: Option<String>,
-}
-
-/// Possible subcommands of the main binary.
-#[derive(Debug, Parser)]
-pub enum Subcommand {
-    /// Key management cli utilities
-    #[clap(subcommand)]
-    Key(KeySubcommand),
-
-    /// The custom benchmark subcommmand benchmarking runtime pallets.
-    #[clap(name = "benchmark", about = "Benchmark runtime pallets.")]
-    Benchmark(frame_benchmarking_cli::BenchmarkCmd),
-
-    /// Try some command against runtime state. Note: `try-runtime` feature must be enabled.
-    #[cfg(feature = "try-runtime")]
-    TryRuntime(try_runtime_cli::TryRuntimeCmd),
-
-    /// Try some command against runtime state. Note: `try-runtime` feature must be enabled.
-    #[cfg(not(feature = "try-runtime"))]
-    TryRuntime,
-
-    /// Verify a signature for a message, provided on STDIN, with a given (public or secret) key.
-    Verify(VerifyCmd),
-
-    /// Generate a seed that provides a vanity address.
-    Vanity(VanityCmd),
-
-    /// Sign a message, with a given (secret) key.
-    Sign(SignCmd),
-
-    /// Build a chain specification.
-    BuildSpec(sc_cli::BuildSpecCmd),
-
-    /// Validate blocks.
-    CheckBlock(sc_cli::CheckBlockCmd),
-
-    /// Export blocks.
-    ExportBlocks(sc_cli::ExportBlocksCmd),
-
-    /// Export the state of a given block into a chain spec.
-    ExportState(sc_cli::ExportStateCmd),
-
-    /// Import blocks.
-    ImportBlocks(sc_cli::ImportBlocksCmd),
-
-    /// Remove the whole chain.
-    PurgeChain(sc_cli::PurgeChainCmd),
-
-    /// Revert the chain to a previous state.
-    Revert(sc_cli::RevertCmd),
-
-    /// Export the genesis state of the parachain.
-    #[clap(name = "export-genesis-state")]
-    ExportGenesisState(ExportGenesisStateCommand),
-
-    /// Export the genesis wasm of the parachain.
-    #[clap(name = "export-genesis-wasm")]
-    ExportGenesisWasm(ExportGenesisWasmCommand),
-}
