--- node/src/cli_relaychain.rs
@@ -1,54 +1,0 @@
-/*
- * This file is part of the Nodle Chain distributed at https://github.com/NodleCode/chain
- * Copyright (C) 2022  Nodle International
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
- */
-use crate::chain_spec;
-use clap::Parser;
-use std::path::PathBuf;
-
-// #[cfg(not(feature = "cli"))]
-#[derive(Debug)]
-pub struct RelayChainCli {
-    /// The actual relay chain cli object.
-    pub base: polkadot_cli::RunCmd,
-
-    /// Optional chain id that should be passed to the relay chain.
-    pub chain_id: Option<String>,
-
-    /// The base path that should be used by the relay chain.
-    pub base_path: Option<PathBuf>,
-}
-
-// #[cfg(not(feature = "cli"))]
-impl RelayChainCli {
-    /// Parse the relay chain CLI parameters using the para chain `Configuration`.
-    pub fn new<'a>(
-        para_config: &sc_service::Configuration,
-        relay_chain_args: impl Iterator<Item = &'a String>,
-    ) -> Self {
-        let extension = chain_spec::Extensions::try_get(&*para_config.chain_spec);
-        let chain_id = extension.map(|e| e.relay_chain.clone());
-        let base_path = para_config
-            .base_path
-            .as_ref()
-            .map(|x| x.path().join("polkadot"));
-        Self {
-            base_path,
-            chain_id,
-            base: polkadot_cli::RunCmd::parse_from(relay_chain_args),
-        }
-    }
-}
