--- node/src/main.rs
@@ -1,33 +1,0 @@
-/*
- * This file is part of the Nodle Chain distributed at https://github.com/NodleCode/chain
- * Copyright (C) 2022  Nodle International
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
- */
-
-mod chain_spec;
-mod cli_relaychain;
-#[macro_use]
-mod service;
-mod cli;
-mod client;
-mod command;
-mod rpc;
-#[macro_use]
-mod para_service;
-mod para_rpc;
-
-fn main() -> sc_cli::Result<()> {
-    command::run()
-}
