+++ runtimes/eden/src/constants.rs
@@ -34,6 +34,8 @@ pub const DOLLARS: Balance = NODL / 100;
 pub const CENTS: Balance = DOLLARS / 100;
 pub const MILLICENTS: Balance = CENTS / 1_000;
 
+pub const EXISTENTIAL_DEPOSIT: Balance = 1 * MILLICENTS;
+
 pub const fn deposit(items: u32, bytes: u32) -> Balance {
     items as Balance * 15 * CENTS + (bytes as Balance) * 6 * CENTS
 }
