+++ runtimes/eden/src/pallets_system.rs
@@ -111,7 +111,7 @@ impl pallet_timestamp::Config for Runtime {
 }
 
 parameter_types! {
-    pub const ExistentialDeposit: Balance = 1 * constants::MILLICENTS;
+    pub const ExistentialDeposit: Balance = constants::EXISTENTIAL_DEPOSIT;
     // For weight estimation, we assume that the most locks on an individual account will be 50.
     // This number may need to be adjusted in the future if this assumption no longer holds true.
     pub const MaxLocks: u32 = 50;
